@@ -1,4 +1,4 @@
-﻿using Aspire.Hosting;
+using Aspire.Hosting;
 
 namespace MyWeatherHub.Tests;
 
@@ -8,18 +8,15 @@ public class EnvVarTests
     [TestMethod]
     public async Task WebResourceEnvVarsResolveToApiService()
     {
-        // Arrange
         var appHost = await DistributedApplicationTestingBuilder
             .CreateAsync<Projects.AppHost>();
 
         var frontend = (IResourceWithEnvironment)appHost.Resources
             .Single(static r => r.Name == "myweatherhub");
 
-        // Act
         var envVars = await frontend.GetEnvironmentVariableValuesAsync(
             DistributedApplicationOperation.Publish);
 
-        // Assert
         CollectionAssert.Contains(envVars,
             new KeyValuePair<string, string>(
                 key: "services__api__https__0",
