@@ -8,7 +8,6 @@ public class IntegrationTests
     [TestMethod]
     public async Task TestApiGetZones()
     {
-        // Arrange
         var appHost = await DistributedApplicationTestingBuilder
             .CreateAsync<Projects.AppHost>();
 
@@ -18,24 +17,16 @@ public class IntegrationTests
         });
 
         await using var app = await appHost.BuildAsync();
-
-        var resourceNotificationService = app.Services
-            .GetRequiredService<ResourceNotificationService>();
-
+        var resourceNotificationService = app.Services.GetRequiredService<ResourceNotificationService>();
         await app.StartAsync();
 
-        // Act
         var httpClient = app.CreateHttpClient("api");
 
-        await resourceNotificationService.WaitForResourceAsync(
-                "api",
-                KnownResourceStates.Running
-            )
+        await resourceNotificationService
+            .WaitForResourceAsync("api", KnownResourceStates.Running)
             .WaitAsync(TimeSpan.FromSeconds(30));
 
         var response = await httpClient.GetAsync("/zones");
-
-        // Assert
         response.EnsureSuccessStatusCode();
         var zones = await response.Content.ReadFromJsonAsync<Zone[]>();
         Assert.IsNotNull(zones);
@@ -45,7 +36,6 @@ public class IntegrationTests
     [TestMethod]
     public async Task TestWebAppHomePage()
     {
-        // Arrange
         var appHost = await DistributedApplicationTestingBuilder
             .CreateAsync<Projects.AppHost>();
 
@@ -55,24 +45,16 @@ public class IntegrationTests
         });
 
         await using var app = await appHost.BuildAsync();
-
-        var resourceNotificationService = app.Services
-            .GetRequiredService<ResourceNotificationService>();
-
+        var resourceNotificationService = app.Services.GetRequiredService<ResourceNotificationService>();
         await app.StartAsync();
 
-        // Act
         var httpClient = app.CreateHttpClient("myweatherhub");
 
-        await resourceNotificationService.WaitForResourceAsync(
-                "myweatherhub",
-                KnownResourceStates.Running
-            )
+        await resourceNotificationService
+            .WaitForResourceAsync("myweatherhub", KnownResourceStates.Running)
             .WaitAsync(TimeSpan.FromSeconds(30));
 
         var response = await httpClient.GetAsync("/");
-
-        // Assert
         response.EnsureSuccessStatusCode();
         var content = await response.Content.ReadAsStringAsync();
         Assert.IsTrue(content.Contains("MyWeatherHub"));
