@@ -1,30 +1,29 @@
 ﻿@page "/"
 @using Microsoft.AspNetCore.Components.QuickGrid
 @using System.Diagnostics
-@using System.Diagnostics.Metrics
 @using Microsoft.EntityFrameworkCore
 @inject NwsManager NwsManager
 @inject ILogger<NwsManager> Logger
 @inject MyWeatherContext DbContext
-@inject ForecastSummarizer Summarizer
 @rendermode InteractiveServer
 
 <PageTitle>My Weather Hub</PageTitle>
 
+<HeadContent>
+	<style>
+
+		.quickgrid {
+			width: 100%;
+		}
+
+		.quickgrid td:has(span.selectedCell) {
+			background-color: yellow;
+		}
+	</style>
+</HeadContent>
+
 <h1>My Weather Hub</h1>
 <h2>Live Weather reports from the US National Weather service</h2>
-
-<!-- Developer Tools Section -->
-<div class="developer-tools-section mb-4">
-	<h5>🛠️ Developer Tools</h5>
-	<div class="btn-group" role="group">
-		<a href="http://localhost:8090" target="_blank" class="btn btn-outline-primary">
-			<i class="bi bi-tools"></i> IT-Tools
-		</a>
-
-	</div>
-
-</div>
 
 <div class="row">
 	<div class="col-md-6">
@@ -34,6 +33,7 @@
 				Show only favorites
 			</label>
 		</div>
+
 		<QuickGrid Items="zones" TGridItem="Zone" Pagination="pagination">
 			<TemplateColumn Title="Name" SortBy="NameSort" Sortable="true">
 				<ColumnOptions>
@@ -42,15 +42,13 @@
 					</div>
 				</ColumnOptions>
 				<ChildContent>
-					<span class="link-primary @(SelectedZone == context ? "selectedCell" : "")" style="cursor: pointer;"
-						@onclick="@(() => SelectZone(context))">@context.Name</span>
+					<span class="link-primary @(SelectedZone == context ? "selectedCell" : "")" style="cursor: pointer;" @onclick="@(() => SelectZone(context))">@context.Name</span>
 				</ChildContent>
 			</TemplateColumn>
 			<TemplateColumn Title="State">
 				<ColumnOptions>
 					<div class="search-box">
-						<input type="search" autofocus @bind="StateFilter" @bind:event="oninput"
-							placeholder="State..." />
+						<input type="search" autofocus @bind="StateFilter" @bind:event="oninput" placeholder="State..." />
 					</div>
 				</ColumnOptions>
 				<ChildContent>
@@ -62,11 +60,11 @@
 					<button @onclick="@(() => ToggleFavorite(context))">
 						@if (FavoriteZones.Contains(context))
 						{
-							<span>&#9733;</span> <!-- Starred -->
+							<span>&#9733;</span>
 						}
 						else
 						{
-							<span>&#9734;</span> <!-- Unstarred -->
+							<span>&#9734;</span>
 						}
 					</button>
 				</ChildContent>
@@ -76,13 +74,9 @@
 	</div>
 
 	<div class="col-md-6" style="max-width:30em;">
-		<p>Click the name of a weather zone from the <a href="https://www.weather.gov/documentation/services-web-api"
-				target="_blank">US National Weather Service</a> and the upcoming weather for that zone will appear
-			below.</p>
-		<p>You can sort and filter the list of available weather forecast zones using the column headers of the grid.
-		</p>
-		<p>Not all forecast zones will have a forecast available. For reliable zones, try Philadelphia, Manhattan,
-			District of Columbia, or Los Angeles County San Gabriel Valley</p>
+		<p>Click the name of a weather zone from the <a href="https://www.weather.gov/documentation/services-web-api" target="_blank">US National Weather Service</a> and the upcoming weather for that zone will appear below.</p>
+		<p>You can sort and filter the list of available weather forecast zones using the column headers of the grid.</p>
+		<p>Not all forecast zones will have a forecast available.  For reliable zones, try Philadelphia, Manhattan, District of Columbia, or Los Angeles County San Gabriel Valley</p>
 		<p>For a zone that isn't reliably reporting, try Bristol Bay, AK (Alaska)</p>
 	</div>
 
@@ -97,28 +91,26 @@
 else if (SelectedZone != null && Forecast != null)
 {
 
-	<div class="forecast-background-container" style="background-image: url('img/@(Summary.ToLowerInvariant())/@(randomBackground).jpg');">
-		<h3 class="weather-headline">
-			Weather for @SelectedZone.Name<text>, </text> @SelectedZone.State (@SelectedZone.Key)
-		</h3>
-		<div class="row row-cols-1 row-cols-md-4 g-4">
-			@foreach (var forecast in Forecast.Take(8))
-			{
-				<div class="col">
-					<div class="card forecast-card">
-						<div class="card-header">@forecast.Name</div>
-						<div class="card-body">@forecast.DetailedForecast</div>
-					</div>
+	<h3>Weather for @SelectedZone.Name<text>, </text> @SelectedZone.State (@SelectedZone.Key)</h3>
+
+	<div class="row row-cols-1 row-cols-md-4 g-4">
+		@foreach (var forecast in Forecast.Take(8))
+		{
+
+			<div class="col">
+				<div class="card">
+					<div class="card-header">@forecast.Name</div>
+					<div class="card-body">@forecast.DetailedForecast</div>
 				</div>
-			}
-		</div>
+			</div>
+
+		}
 	</div>
 
 }
 else if (SelectedZone != null && !string.IsNullOrEmpty(Error))
 {
-	<div class="alert-danger border border-1 border-danger p-3" style="width: 30em">@Error<br />Choose another weather
-		station</div>
+	<div class="alert-danger border border-1 border-danger p-3" style="width: 30em">@Error<br />Choose another weather station</div>
 }
 
 @code {
@@ -135,10 +127,10 @@ else if (SelectedZone != null && !string.IsNullOrEmpty(Error))
 			}
 
 			results = string.IsNullOrEmpty(StateFilter) ? results.AsQueryable()
-			: results.Where(z => z.State == StateFilter.ToUpper()).AsQueryable();
+				: results.Where(z => z.State == StateFilter.ToUpper()).AsQueryable();
 
 			results = string.IsNullOrEmpty(NameFilter) ? results
-			: results.Where(z => z.Name.Contains(NameFilter, StringComparison.InvariantCultureIgnoreCase));
+				: results.Where(z => z.Name.Contains(NameFilter, StringComparison.InvariantCultureIgnoreCase));
 
 			return results.OrderBy(z => z.Name);
 		}
@@ -153,7 +145,7 @@ else if (SelectedZone != null && !string.IsNullOrEmpty(Error))
 	string StateFilter { get; set; } = string.Empty;
 
 	GridSort<Zone> NameSort = GridSort<Zone>
-	.ByAscending(f => f.Name);
+			.ByAscending(f => f.Name);
 
 	Zone SelectedZone { get; set; } = null!;
 
@@ -164,12 +156,7 @@ else if (SelectedZone != null && !string.IsNullOrEmpty(Error))
 	bool IsLoading = false;
 
 	bool ShowOnlyFavorites { get; set; }
-
 	List<Zone> FavoriteZones { get; set; } = new List<Zone>();
-
-	string Summary { get; set; } = string.Empty;
-
-	int randomBackground = new Random().Next(1, 4);
 
 	protected override async Task OnInitializedAsync()
 	{
@@ -179,6 +166,10 @@ else if (SelectedZone != null && !string.IsNullOrEmpty(Error))
 
 	private async Task SelectZone(Zone zone)
 	{
+		// Workaround to create a new trace
+		// Blazor server keeps connection open so a new trace isn't created
+		Activity.Current = null;
+
 		SelectedZone = zone;
 		IsLoading = true;
 		StateHasChanged();
@@ -198,10 +189,6 @@ else if (SelectedZone != null && !string.IsNullOrEmpty(Error))
 			Error = $"Unable to locate weather for {SelectedZone.Name}({SelectedZone.Key})";
 		}
 
-		if (string.IsNullOrEmpty(Error))
-		{
-			Summary = await Summarizer.SummarizeForecastAsync(Forecast.FirstOrDefault().DetailedForecast);
-		} 
 	}
 
 	private async Task ToggleFavorite(Zone zone)
@@ -218,5 +205,6 @@ else if (SelectedZone != null && !string.IsNullOrEmpty(Error))
 		}
 		await DbContext.SaveChangesAsync();
 	}
+
 
 }
